requirements.txt
# --------- pytorch --------- #
# torch>=1.11.0
# torchvision>=0.12.0
pytorch-lightning==1.7.7
torchmetrics==0.11.0
# torchdata
timm==0.6.12
einops
six

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
sh              # for running bash commands in some tests
tqdm
h5py
click
netcdf4
xarray
dask